+++ Program.cs
@@ -11,6 +11,7 @@ namespace Teron_Addon_Manager
             // To customize application configuration such as set high DPI settings or default font,
             // see https://aka.ms/applicationconfiguration.
             ApplicationConfiguration.Initialize();
+            Application.SetDefaultFont(new Font("Segoe UI", 10F));
             Application.Run(new Addon_Manager());
         }
     }